#Refactory fo Handler part2

+++ src/Application/Features/Styles/Commands/DeleteTagInStyle.cs
@@ -4,7 +4,9 @@ using Application.Extensions;
 using Domain.Entities.MidjourneyStyles;
 using Domain.ValueObjects;
 using FluentResults;
+using Domain.Errors;
 using static Application.Errors.ApplicationErrorMessages;
+using static Domain.Errors.DomainErrorMessages;
 
 namespace Application.Features.Styles.Commands.RemoveTagInStyle;
 
@@ -18,11 +20,25 @@ public static class DeleteTagInStyle
 
         public async Task<Result<MidjourneyStyle>> Handle(Command command, CancellationToken cancellationToken)
         {
+            List<DomainError> domainErrors = [];
+
+            domainErrors
+                .CollectErrors<StyleName>(command.StyleName)
+                .CollectErrors<Tag>(command.Tag);
+
             List<ApplicationError> applicationErrors = [];
 
             applicationErrors
                 .IfStyleNotExists(command.StyleName, _styleRepository)
                 .IfTagNotExists(command.StyleName, command.Tag, _styleRepository);
+
+            if (applicationErrors.Count != 0 || domainErrors.Count != 0)
+            {
+                var error = new Error("Validation failed")
+                    .WithMetadata("Application Errors", applicationErrors)
+                    .WithMetadata("Domain Errors", domainErrors);
+                return Result.Fail<MidjourneyStyle>(error);
+            }
 
             return await _styleRepository.DeleteTagFromStyleAsync(command.StyleName, command.Tag);
         }
